Adds multi-select
