+++ Assets/Scripts/LeanScripts/LeanSelect.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: ad7dc99ce51e599479fc327e22a8de27
-timeCreated: 1513139858
-licenseType: Store
+guid: 4f83827ec28a73b40835c871807d982a
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
