+++ Assets/Scripts/LeanScripts/LeanSelectionBox.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 479d695282ba6cf4ba3ab3c4848f05e8
-timeCreated: 1480382982
-licenseType: Store
+guid: ec0dd9a6eb6531647b93d8fa2436e936
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
