+++ Assets/Scripts/Managers/ContraptionsManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d744fd6c45d03534aa12476772fa9be1
+guid: b8b9c792fd29f4945927246cc3b0d1d0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
