+++ Assets/Scripts/Managers/PartsManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a99c25dcb8eabaf41afc854416ebd1a0
+guid: 675e9270413c47c48ad3ff46a8ae0a78
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
